+++ sdk/redhatopenshifthcp/azure-resourcemanager-redhatopenshifthcp/src/main/java/com/azure/resourcemanager/redhatopenshifthcp/fluent/HcpOpenShiftClustersClient.java
@@ -13,6 +13,7 @@ import com.azure.core.util.Context;
 import com.azure.core.util.polling.SyncPoller;
 import com.azure.resourcemanager.redhatopenshifthcp.fluent.models.HcpOpenShiftClusterAdminCredentialInner;
 import com.azure.resourcemanager.redhatopenshifthcp.fluent.models.HcpOpenShiftClusterInner;
+import com.azure.resourcemanager.redhatopenshifthcp.models.HcpOpenShiftClusterAdminCredentialRequest;
 
 /**
  * An instance of this class provides access to all the operations defined in HcpOpenShiftClustersClient.
@@ -275,6 +276,7 @@ public interface HcpOpenShiftClustersClient {
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param hcpOpenShiftClusterName The name of the HcpOpenShiftCluster.
+     * @param body The content of the action request.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
@@ -282,13 +284,15 @@ public interface HcpOpenShiftClustersClient {
      */
     @ServiceMethod(returns = ReturnType.LONG_RUNNING_OPERATION)
     SyncPoller<PollResult<HcpOpenShiftClusterAdminCredentialInner>, HcpOpenShiftClusterAdminCredentialInner>
-        beginRequestAdminCredential(String resourceGroupName, String hcpOpenShiftClusterName);
+        beginRequestAdminCredential(String resourceGroupName, String hcpOpenShiftClusterName,
+            HcpOpenShiftClusterAdminCredentialRequest body);
 
     /**
      * Request a temporary admin kubeconfig for the cluster.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param hcpOpenShiftClusterName The name of the HcpOpenShiftCluster.
+     * @param body The content of the action request.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
@@ -297,13 +301,15 @@ public interface HcpOpenShiftClustersClient {
      */
     @ServiceMethod(returns = ReturnType.LONG_RUNNING_OPERATION)
     SyncPoller<PollResult<HcpOpenShiftClusterAdminCredentialInner>, HcpOpenShiftClusterAdminCredentialInner>
-        beginRequestAdminCredential(String resourceGroupName, String hcpOpenShiftClusterName, Context context);
+        beginRequestAdminCredential(String resourceGroupName, String hcpOpenShiftClusterName,
+            HcpOpenShiftClusterAdminCredentialRequest body, Context context);
 
     /**
      * Request a temporary admin kubeconfig for the cluster.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param hcpOpenShiftClusterName The name of the HcpOpenShiftCluster.
+     * @param body The content of the action request.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
@@ -311,13 +317,14 @@ public interface HcpOpenShiftClustersClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     HcpOpenShiftClusterAdminCredentialInner requestAdminCredential(String resourceGroupName,
-        String hcpOpenShiftClusterName);
+        String hcpOpenShiftClusterName, HcpOpenShiftClusterAdminCredentialRequest body);
 
     /**
      * Request a temporary admin kubeconfig for the cluster.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param hcpOpenShiftClusterName The name of the HcpOpenShiftCluster.
+     * @param body The content of the action request.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
@@ -326,7 +333,7 @@ public interface HcpOpenShiftClustersClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     HcpOpenShiftClusterAdminCredentialInner requestAdminCredential(String resourceGroupName,
-        String hcpOpenShiftClusterName, Context context);
+        String hcpOpenShiftClusterName, HcpOpenShiftClusterAdminCredentialRequest body, Context context);
 
     /**
      * Revoke all credentials issued by requestAdminCredential.