+++ sdk/redhatopenshifthcp/azure-resourcemanager-redhatopenshifthcp/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.0.0-beta.1 (2026-08-19)
 
+- Azure Resource Manager Red Hat OpenShift Hosted Control Planes client library for Java. This package contains Microsoft Azure SDK for Red Hat OpenShift Hosted Control Planes Management SDK.  Package api-version 2026-09-01-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 - Azure Resource Manager Red Hat OpenShift Hosted Control Planes client library for Java. This package contains Microsoft Azure SDK for Red Hat OpenShift Hosted Control Planes Management SDK.  Package api-version 2026-06-30-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 ### Features Added
 
 - Initial release for the azure-resourcemanager-redhatopenshifthcp Java SDK.
+